* Added doc/README.developers, these are important instructions for developers on the project
* Modified config.cpp so that the debugging output of the
configuration file parsing is only compiled with DEBUG_EXTENDED turned
on - the amount of data posted was more appropriate for DEBUG_EXTENDED
in this case.

Also edited the autoconf details, a minor change. the line 'use make
or gmake' was slightly inaccurate now that the makefiles no longer
rely on GNU make specifically. If I'm wrong it would be better that
the makefiles are fixed to be more compatible rather than just rely on
people using GNU make.