Started moving DEFAULT_CONFIG_* defines from autoconf.h (where they are just bloat to everything but one file which includes it - urgh) into the new src/config/configdefault.h. The defines are being renamed to remove the CONFIG_ bit in the middle (which is a bit of a waste of space). This means only the src/config/vars.cpp file can include the defaults which will hopefully end up shrinking the size of the autoconf.h file down to what it should be - stuff which gets configured by the configuration script rather than static defines.
The end result also means the user who compiles the source will have a
harder time finding where the definitions are. Sure, this is mean, but the
other side (the good side) is that it will (hopefully) encourage the user to
edit their configuration file PROPERLY rather than hack the code to
configure their server in a dodgey way. If they do need to edit the default
defines, then there is perhaps a design flaw which should probably be
corrected, but this is doubtful since the configuration file is being
designed to be flexible/powerful enough to make even the biggest tweak-freak
happy. Perhaps I should test this theory out on Scott when the Config class
and parser are complete before presuming that to be the case :)