Removed the mandatory configuration flag, the modules will have to deal with it on their own.. consider the following:
module ./foo.so;
module ./foo.so {};

If foo.so had that flag turned on, the first would cause an error but the
second wouldn't. Without majorly changing the way the config parser.. well..
parses (such as stopping it from being nice and passive as it currently is,
which IMHO is one of its strengths) and without adding a mess of variables,
(booleans, or whatever) to check later it is simply better for the modules
to check what it is they need to be configured.

The best way for them to do this IMHO is for them to check in their start
function (when the configuration would have definately by then finished
loading) and then complain about what THEY needed to be configured. They can
then return false back to the daemon after logging their complaint, or
whatever, and the daemon can unload the module and leave it in the user's
hands.