Added much requested support for prefixing oper generated wallops with a character.
The implementation is fairly flexible, actually being a langtags tag which
acts as a wrapper (so it's not just a prefix char) for the message, allowing
for who knows what (I know this will make one person happy :).

Currently the language file is set up to simply prefix the messages in the
same style as ircu on (undernet) -- there's no language specific data so
there's no need to 'translate' it.

It has been suggested, though, that instead of a dollar sign (U+0024)
prefix, a place-of-interest sign should be used (U+2318) in order to make
it stand out. Perhaps this is something to look into in the future, but for
now it will probably help script kiddies more with something in the low
7-bit UCS range (ie. ASCII).